comit

+++ resources/views/admin/lifestyle/scriptjs.blade.php
@@ -123,37 +123,70 @@
             var userURL = $(this).data("url");
 
             event.preventDefault();
-            swal({
-                title: `Are you sure you want to delete this record?`,
-                text: "If you delete this, it will be gone forever.",
-                icon: "warning",
-                buttons: true,
-                dangerMode: true,
-            }).then((willDelete) => {
-                if (willDelete) {
 
-                    $.ajax({
-                        headers: {
-                            "X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr(
-                                "content"
-                            ),
-                        },
-                        method: "DELETE",
-                        url: userURL,
-                        dataType: "json",
-                        success: function(output) {
-                            if (output == true) {
-                                sessionStorage.setItem('showSuccessNotification', 'true');
-                                location.reload();
+            $.ajax({
+                headers: {
+                    "X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr(
+                        "content"
+                    ),
+                },
+                method: "post",
+                url: "{{route('lifestyle.selectedbyuser')}}",
+                data: {
 
-                            } else {
-                                sessionStorage.setItem('showErrorNotification', 'true');
-                                toastr.error("Lifestyle don't Deleted !");
+                    id: hobbyId
+                },
+                dataType: "json",
+                success: function(output) {
+                    if (output == false) {
+                        swal({
+                            title: `Error`,
+                            text: "User(s) belongs to deleted Interest And Hobby. can not delete Interest And Hobby",
+                            icon: "error",
+
+                        })
+
+                    } else {
+                        swal({
+                            title: `Are you sure you want to delete this record?`,
+                            text: "If you delete this, it will be gone forever.",
+                            icon: "warning",
+                            buttons: true,
+                            dangerMode: true,
+                        }).then((willDelete) => {
+                            if (willDelete) {
+
+                                $.ajax({
+                                    headers: {
+                                        "X-CSRF-TOKEN": $('meta[name="csrf-token"]').attr(
+                                            "content"
+                                        ),
+                                    },
+                                    method: "DELETE",
+                                    url: userURL,
+                                    dataType: "json",
+                                    success: function(output) {
+                                        if (output == true) {
+                                            sessionStorage.setItem('showSuccessNotification', 'true');
+                                            location.reload();
+
+                                        } else {
+                                            swal({
+                                                title: `Error`,
+                                                text: "User(s) belongs to deleted Interest And Hobby. can not delete Interest And Hobby",
+                                                icon: "error",
+
+                                            })
+                                        }
+                                    },
+                                });
                             }
-                        },
-                    });
+                        });
+                    }
                 }
             });
+
+
         });
 
         if (sessionStorage.getItem('showSuccessNotification')) {
